Change: Pass the fn updateMonthYEar into useDate
